+++ src/main/java/tech/habegger/elastic/shared/BoundsSpec.java
@@ -1,6 +1,6 @@
 package tech.habegger.elastic.shared;
 
 public record BoundsSpec(
-    float min,
-    float max
+    double min,
+    double max
 ) { }
