+++ src/main/java/tech/habegger/elastic/shared/Range.java
@@ -3,16 +3,16 @@ package tech.habegger.elastic.shared;
 import com.fasterxml.jackson.annotation.JsonInclude;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public record Range(Float from, Float to, String key) {
-    public static Range between(Float from, Float to) {
+public record Range(Double from, Double to, String key) {
+    public static Range between(Double from, Double to) {
         return new Range(from, to, null);
     }
 
-    public static Range from(Float from) {
+    public static Range from(Double from) {
         return new Range(from, null, null);
     }
 
-    public static Range to(Float to) {
+    public static Range to(Double to) {
         return new Range(null, to, null);
     }
